requirements.txt
cupy_cuda115==10.6.0
matplotlib==3.7.1
natsort==8.2.0
nibabel==4.0.2
numpy==1.24.3
Requests==2.31.0
scipy==1.9.1
torch==2.1.1
torchvision==0.16.1
tqdm==4.63.1